+++ Connector.php
@@ -21,15 +21,21 @@ class Shopware_Controllers_Frontend_Jtlconnector extends Enlight_Controller_Acti
     {
         session_destroy();
         define('CONNECTOR_DIR', __DIR__);
-
+    
+        // Tmp directory fallback
+        $dir = sys_get_temp_dir();
+        if (!is_writeable($dir)) {
+            $dir = CONNECTOR_DIR . DIRECTORY_SEPARATOR . 'tmp';
+        }
+        
         $application = null;
 
         try {
             if (file_exists(CONNECTOR_DIR . '/connector.phar')) {
-                if (is_writable(sys_get_temp_dir())) {
+                if (is_writable($dir)) {
                     include_once('phar://' . CONNECTOR_DIR . '/connector.phar/src/bootstrap.php');
                 } else {
-                    echo sprintf('Directory %s is not writeable. Please contact your administrator or hoster.', sys_get_temp_dir());
+                    echo sprintf('Directory %s is not writeable. Please contact your administrator or hoster.', $dir);
                 }
             } else {
                 include_once(CONNECTOR_DIR . '/src/bootstrap.php');