+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 -------------------
 - Added missing product specific group fail safe
 - Added customer order pull start date (via config parameter customer_order_pull_start_date)
+- Fixed customer number
+- Fixed category meta title in non default languages
 
 2.0.13
 ------