+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added customer order pull start date (via config parameter customer_order_pull_start_date)
 - Fixed customer number
 - Fixed category meta title in non default languages
+- Fixed product filter group allocation after set removal
 
 2.0.13
 ------