+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-2.0.14 (unreleased)
--------------------
+2.0.14
+------
 - Added missing product specific group fail safe
 - Added customer order pull start date (via config parameter customer_order_pull_start_date)
 - Added new category attribute management support